requirements.txt
pandas==1.3.3
pymysql==1.0.2
requests==2.26.0
SQLAlchemy==1.4.25
python-dotenv==0.19.1